requirement.txt
apache-airflow==2.2.2
pandas==1.3.0
numpy==1.21.1
msgpack==1.0.3
lz4==3.1.10
jupyter-server-proxy
distributed==2021.6.0
blosc==1.10.2
dask==2021.6.0